[FIX] hr_holidays_public: Traceback when change the date was fixed
The field date in the public holidays cannot be changed when the field
to block the field cancellation is True.

With the attrs to block the field, if the record is not saved, the date
field is clear and do not allow to save the record because is a required
field. Then was improved with the onchange, to show a error message if
the date is changed and the check to block is True.

Closes #569